pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "crowd-management-system"
version = "0.1.0"
authors = [{ name = "ncomputers" }]
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn",
    "redis",
    "opencv-python-headless",
    "torch",
    "onnxruntime-gpu",
    "ultralytics",
    "deep-sort-realtime",
    "loguru",
    "jinja2",
    "openpyxl",
    "websockets",
    "python-multipart",
    "imagehash",
    "Pillow",
    "qrcode[pil]",
    "passlib",
    "pytest",
    "setuptools<81",
    "psutil",
    "numpy",
    "authlib",
    "weasyprint",
    "httpx",
    "email-validator",
    "itsdangerous",
    "pydantic-settings",
    "fastapi-csrf-protect",
    "ffmpeg-python",
    "keyring",
    "fakeredis",
    "PyYAML",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["app", "modules", "utils", "services"]

[tool.ruff]
line-length = 100
fix = true

[tool.ruff.lint]
extend-select = ["B", "C4", "Q", "T20"]
ignore = [
    "E203",
    "E266",
    "E501",
    "E402",
    "B008",
    "T201",
    "B904",
    "F821",
    "E731",
    "B023",
    "E711",
    "F401",
    "F811",
    "E741",
    "F841",
    "C405",
]

[tool.pytest.ini_options]
addopts = "-ra -k 'not test_faces and not test_tracker_redis and not test_admin_user_management'"
markers = [
    "slow: marks tests as slow",
    "integration: marks tests that require external services",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore:Specified provider 'CUDAExecutionProvider':UserWarning",
]
testpaths = ["tests"]
xfail_strict = true

[tool.coverage.run]
branch = true
source = ["app", "modules", "utils", "services"]

[tool.coverage.report]
show_missing = true
skip_covered = true
[tool.setuptools.packages.find]
where = ["."]
include = ["core", "modules", "routers", "schemas", "utils"]
